Add edit, remove offer functionality
