+++ src/pages/OffersPage/components/OffersTabContent/components/OfferCard/OfferCard.module.less
@@ -52,6 +52,10 @@
   svg {
     height: 12px !important;
     width: 12px !important;
+
+    rect {
+      fill: var(--content-primary);
+    }
   }
 }
 .editOfferButton,
