+++ src/pages/OffersPage/components/OffersTabContent/components/OfferCard/hooks.ts
@@ -1,66 +1,18 @@
-import { FC, useMemo } from 'react'
+import { useMemo } from 'react'
 
 import { useConnection, useWallet } from '@solana/wallet-adapter-react'
 import { useNavigate } from 'react-router-dom'
 import { TxnExecutor } from 'solana-transactions-executor'
 
-import { Button } from '@banx/components/Buttons'
-
-import { Offer } from '@banx/api/core'
+import { CollectionMeta, Offer } from '@banx/api/core'
 import { useUserOffers } from '@banx/pages/OffersPage/hooks'
 import { PATHS } from '@banx/router'
 import { useMarketsURLControl, useSyntheticOffers } from '@banx/store'
 import { defaultTxnErrorHandler } from '@banx/transactions'
 import { makeRemoveOfferAction } from '@banx/transactions/bonds'
-import { enqueueSnackbar, trackPageEvent } from '@banx/utils'
+import { enqueueSnackbar } from '@banx/utils'
 
-import { TableUserOfferData } from '../helpers'
-
-import styles from '../PendingOffersTable.module.less'
-
-interface ActionsCellProps {
-  offer: TableUserOfferData
-  isCardView: boolean
-}
-
-export const ActionsCell: FC<ActionsCellProps> = ({ offer, isCardView }) => {
-  const { removeOffer, goToEditOffer } = useActionsCell(offer)
-
-  const buttonSize = isCardView ? 'medium' : 'small'
-
-  const onEdit = () => {
-    goToEditOffer()
-    trackPageEvent('myoffers', 'pendingtab-edit')
-  }
-
-  const onRemove = () => {
-    removeOffer()
-    trackPageEvent('myoffers', 'pendingtab-remove')
-  }
-
-  return (
-    <div className={styles.actionsButtons}>
-      <Button
-        className={styles.actionButton}
-        onClick={onEdit}
-        variant="secondary"
-        size={buttonSize}
-      >
-        Edit
-      </Button>
-      <Button
-        className={styles.removeButton}
-        onClick={onRemove}
-        variant="secondary"
-        size={buttonSize}
-      >
-        Remove
-      </Button>
-    </div>
-  )
-}
-
-const useActionsCell = (offer: TableUserOfferData) => {
+export const useActionsCell = (offer: Offer, collectionMeta: CollectionMeta) => {
   const wallet = useWallet()
   const { connection } = useConnection()
   const { offers, updateOrAddOffer } = useUserOffers()
@@ -81,15 +33,15 @@ const useActionsCell = (offer: TableUserOfferData) => {
     setSyntheticOffer({
       isEdit: true,
       publicKey: offer.publicKey,
-      loanValue: offer.loanValue,
-      loansAmount: offer.loansAmount,
+      loanValue: offer.currentSpotPrice,
+      loansAmount: offer.buyOrdersQuantity,
       assetReceiver: offer.assetReceiver,
       marketPubkey: offer.hadoMarket,
       mathCounter: 0,
       deltaValue: 0,
     })
 
-    const collectionName = offer.collectionName
+    const collectionName = collectionMeta.collectionName
 
     setMarketVisibility(collectionName, true)
     setSelectedMarkets([collectionName])
