+++ src/pages/OffersPage/components/OffersTabContent/components/Summary/Summary.module.less
@@ -68,7 +68,7 @@
   }
 }
 
-.terminateButton {
+.terminateButton:not(:disabled) {
   background: var(--additional-lava-secondary) !important;
   border-color: var(--additional-lava-primary) !important;
 }
