+++ src/pages/OffersPage/components/PendingOffersTable/TableCells/index.ts
@@ -1,4 +1,3 @@
-export * from './ActionsCell'
 export * from './APRCell'
 export * from './InterestCell'
 export * from './OfferCell'
